add a macro for making configuration options slightly less error-prone (#63)

+++ src/log/periodic_flusher.rs
@@ -9,14 +9,14 @@ use super::*;
 pub fn flusher(name: String,
                iob: Arc<IoBufs>,
                shutdown: Arc<AtomicBool>,
-               flush_every_ms: u64)
+               flush_every_ms: usize)
                -> std::io::Result<thread::JoinHandle<()>> {
     thread::Builder::new()
         .name(name)
         .spawn(move || {
             while !shutdown.load(SeqCst) {
                 iob.flush();
-                thread::sleep(Duration::from_millis(flush_every_ms));
+                thread::sleep(Duration::from_millis(flush_every_ms as u64));
             }
         })
 }
